Look in certain hardcoded paths for Vagrant binary
@homebrew installs vagrant at `/usr/local/bin`, which isn't in the 
default $PATH in the OS X bash.

E.g. if I have a certain $PATH setup in zsh or fish, the vagrant binary
will work in the terminal but not in Vagrant Bar.

This is not the perfect fix, but I think it will fix the problem for
most people.